Move history event sanitization out of cmd/main
